requirements.txt
gradio==4.44.0
python-dotenv
